20201115 작업 push, 미완성

+++ src/main/java/com/hoondragonite/aboutme/repository/CareerinfoRepository.java
@@ -3,5 +3,5 @@ package com.hoondragonite.aboutme.repository;
 import com.hoondragonite.aboutme.domain.careerinfo.Careerinfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CareerinfoRepository extends JpaRepository<Careerinfo, Long> {
+public interface CareerinfoRepository extends JpaRepository<Careerinfo, Integer> {
 }
